requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
future==0.17.1
idna==2.8
more-itertools==7.0.0
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
pluggy==0.9.0
py==1.8.0
pytest==4.4.0
pytest-cov==2.6.1
python-dateutil==2.8.0
python-twitter==3.5
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
urllib3==1.24.1
